requirements.txt
absl-py==1.4.0
altgraph==0.17.3
asgiref==3.6.0
attrs==23.1.0
autopep8==2.0.2
beautifulsoup4==4.12.2
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cloudinary==1.34.0
contourpy==1.0.7
crispy-bootstrap5==0.7
cryptography==39.0.1
cvzone==1.5.6
cycler==0.11.0
defusedxml==0.7.1
Django==4.1.5
django-allauth==0.52.0
django-appconf==1.0.5
django-cors-headers==3.13.0
django-crispy-forms==2.0
django-debug-toolbar==4.0.0
django-environ==0.9.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
EasyProcess==1.1
Eel==0.16.0
entrypoint2==1.1
exceptiongroup==1.1.1
fire==0.5.0
flatbuffers==23.5.9
fonttools==4.39.3
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-resources==5.12.0
iniconfig==2.0.0
jeepney==0.8.0
kiwisolver==1.4.4
libxml2-python==2.11.4
matplotlib==3.7.1
mediapipe==0.10.0
MouseInfo==0.1.3
mss==9.0.1
numpy==1.24.3
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.0
Pillow==10.0.0
pluggy==1.0.0
protobuf==3.20.3
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg2-binary==2.9.6
PyAutoGUI==0.9.54
pycodestyle==2.10.0
pycparser==2.21
PyGetWindow==0.0.9
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
PyJWT==1.7.1
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
pytest==7.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
python3-xlib==0.15
pytweening==1.0.7
pytz==2022.7.1
rcssmin==1.1.1
requests==2.28.2
requests-oauthlib==1.3.1
rjsmin==1.2.1
six==1.16.0
sounddevice==0.4.6
soupsieve==2.4.1
sqlparse==0.4.3
termcolor==2.2.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==1.26.14
whichcraft==0.6.1
whitenoise==6.3.0
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
